[JENKINS-21202] - Generalization of Command installers This change adds an AbstractCommandInstaller, which can be extended for every script-based installation command.
Related to https://issues.jenkins-ci.org/browse/JENKINS-21202
