+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/config.jelly
