+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/config_fr.properties
