+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/config_ja.properties
