+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/config_nl.properties
