+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/config_pt_BR.properties
