+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/config_zh_TW.properties
