+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help-command.html
