+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help-command_zh_TW.html
