+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help-toolHome_de.html
