+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help-toolHome_zh_CN.html
