+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help.html
