+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help_de.html
