+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help_ja.html
