+++ core/src/main/resources/hudson/tools/AbstractCommandInstaller/help_zh_TW.html
